requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2025.6.15
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
decorator==5.2.1
defusedxml==0.7.1
deptry==0.23.0
discord.py==2.4.0
docopt==0.6.2
executing==2.2.0
fastapi==0.115.13
fastjsonschema==2.21.1
feedparser==6.0.11
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
jedi==0.19.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
multidict==6.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
requirements-parser==0.13.0
rpds-py==0.25.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.46.2
tinycss2==1.4.0
tornado==6.5.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.2.3
uvicorn==0.34.3
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
yarg==0.1.9
yarl==1.20.1
